Fixed endpoint-column extraction bug

+++ src/Engine/Cloud/AletheiaCloudClient.cs
@@ -1827,7 +1827,7 @@ namespace Aletheia.Engine.Cloud
             //Endpoint
             try
             {
-                ToReturn.Endpoint = (AletheiaEndpoint)dr.GetInt32(dr.GetOrdinal(prefix + "Endpoint"));
+                ToReturn.Endpoint = (AletheiaEndpoint)dr.GetByte(dr.GetOrdinal(prefix + "Endpoint"));
             }
             catch
             {
